moviedb-api/requirements.txt
Django>=5.0.7
requests>=2.31.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
